[15m] saco el armado de json a otro archivo
